Moving nodes type in another file
